% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp_e0}
\alias{decomp_e0}
\title{decompose life expectancy using a variety of methods}
\usage{
decomp_e0(
  mx1,
  mx2,
  age,
  sex1,
  sex2,
  method = c("ltre_lt", "ltre_arriaga_instantaneous", "ltre_numerical", "arriaga",
    "arriaga_symmetrical", "horiuchi"),
  N = 20,
  closeout = TRUE,
  opt = TRUE,
  tol = 1e-10,
  ...
)
}
\arguments{
\item{mx1}{numeric. age-structured mortality rates for population 1}

\item{mx2}{numeric. age-structured mortality rates for population 2, must be same length as \code{mx1}}

\item{age}{integer. lower bound of each age group, must be same length as \code{mx1}}

\item{sex1}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{sex2}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{method}{character. \code{"ltre_lt"},\code{"ltre_arriaga_instantaneous"},\code{"ltre_numerical"},\code{"arriaga"}, \code{"arriaga_symmetrical"}, or \code{"horiuchi"}}

\item{N}{integer. used for LTRE or Horiuchi methods}

\item{closeout}{logical. Do we handle closeout, or truncate at top age_}

\item{...}{optional arguments passed to \code{numDeriv::grad()}}
}
\description{
A variety of exact or sympototically exact life expectancy decomposition methods are implemented. These include the lifetable response experiment, using three potential sensitivity functions (lifetable-based \code{ltre_lt}, an instantaneous Arriaga-derived sensitivity \code{ltre_arriaga_instantaneous}, or numerical sensitivity \code{ltre_numerical}), the directional Arriaga method \code{arriaga}, a symmetrical Arriaga method \code{arriaga_symmetrical}, or the linear intregal (a.k.a pseudo continuous) method of Horiuchi et al \code{horiuchi}. These all give similar results, except for directional \code{arriaga}, which is the most different.
}
