% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{LEdecomp}
\alias{LEdecomp}
\title{decompose life expectancy using a variety of methods}
\usage{
LEdecomp(
  mx1,
  mx2,
  age,
  sex1 = "t",
  sex2 = sex1,
  method = c("lifetable", "arriaga_instantaneous", "arriaga_instantaneous2", "numerical",
    "arriaga", "arriaga_symmetrical"),
  closeout = TRUE,
  opt = TRUE,
  tol = 1e-10,
  extras = FALSE,
  ...
)
}
\arguments{
\item{mx1}{numeric. age-structured mortality rates for population 1}

\item{mx2}{numeric. age-structured mortality rates for population 2, must be same length as \code{mx1}}

\item{age}{integer. lower bound of each age group, must be same length as \code{mx1}}

\item{sex1}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{sex2}{character. \code{"m"},\code{"f"}, or \code{"t"}, affects a0 treatment.}

\item{method}{character. \code{"lifetable"},\code{"arriaga_instantaneous"},\code{"arriaga_instantaneous2"},\code{"numerical"},\code{"arriaga"}, \code{"arriaga_symmetrical"}}

\item{closeout}{logical. Do we handle closeout, or truncate at top age_}

\item{opt}{logical, default \code{TRUE}. For sensitivity-based decompositions, shall we optimize the rate averaging to eliminate the decomposition residual?}

\item{tol}{numeric, default \code{1e-10}, tolerance parameter for rate averaging optimization.}

\item{...}{optional arguments passed to \code{numDeriv::grad()}}
}
\description{
A variety of exact or asympototically exact life expectancy decomposition methods are implemented. These include the lifetable response experiment, using three potential sensitivity functions (lifetable-based \code{ltre_lt}, an instantaneous Arriaga-derived sensitivity \code{ltre_arriaga_instantaneous}, or numerical sensitivity \code{ltre_numerical}), the directional Arriaga method \code{arriaga}, a symmetrical Arriaga method \code{arriaga_symmetrical}. These all give similar results, except for directional \code{arriaga}, which is the most different.
}
\details{
LEdecomp
}
