% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_utils.R
\name{lL_to_ex}
\alias{lL_to_ex}
\title{calculate remaining life expectancy \code{ex} for each age}
\usage{
lL_to_ex(lx, Lx)
}
\arguments{
\item{lx}{numeric vector of lifetable survivors at exact age \code{x}}

\item{Lx}{numeric vector of lifetable exposure \code{Lx}}
}
\value{
numeric vector of remaining life expectancy \code{ex}
}
\description{
Here we combine HMD Method Protocol equations 79 and 80. We calculate all the remaining years left to live at each age, then condition this on survival to each age.
}
\references{
\insertRef{wilmoth2021methods}{coddecomp}
}
