% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_utils.R
\name{lx_to_dx}
\alias{lx_to_dx}
\title{Calculate the lifetable death distribution}
\usage{
lx_to_dx(lx)
}
\arguments{
\item{lx}{numeric vector of lifetable survivors at exact age \code{x}}
}
\value{
numeric vector of \code{dx} values
}
\description{
Minus the decumulation of the survival curve gives the death distribution. Or the element-wise product of \code{lx} and the conditional death probabilities \code{qx} gives the same thing.
}
\seealso{
\link[DemoTools]{lt_id_l_d}
}
