% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_utils.R
\name{mx_to_qx}
\alias{mx_to_qx}
\title{produce single-age \code{qx} values}
\usage{
mx_to_qx(mx, ax, closeout = TRUE)
}
\arguments{
\item{mx}{numeric vector of the mortality rates (central death rates)}

\item{ax}{numeric vector of \code{ax} values}

\item{closeout}{logical. Default \code{TRUE}.}
}
\description{
\code{qx} gives conditional death probabilities, in this case forced to be consistent with a set of \code{mx} and \code{ax} values per HMD Method Protocol eq 71.
}
\references{
\insertRef{wilmoth2021methods}{coddecomp}
}
\seealso{
\link[DemoTools]{lt_id_ma_q}
}
