% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_utils.R
\name{rcumsum}
\alias{rcumsum}
\title{top-down cumulative sums, as in the lifetable Tx}
\usage{
rcumsum(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector the same length as \code{x}
}
\description{
Why write \code{x |> rev() |> cumsum() |> rev()} when you can just write \code{rcumsum(x)}?
}
